Mypage 리팩토링
