+++ src/Mypage/SigninPage.scss
@@ -1,0 +1,45 @@
+@import '../styles/variables.scss';
+
+
+.existingCustomers {
+  width: 50%;
+  height: 300px;
+  position: relative;
+  
+  h2 {
+    font-size: 20px;
+    font-weight: bold;
+    margin:0 0 20px;
+  }
+  
+  .inputText {
+    display: block;
+    width: 100%; 
+    height: 40px;
+    padding: 15px;
+    font-size: 15px;
+    margin-bottom: 20px;
+    border: 1px solid #ccc;
+    outline: none;
+  }
+  
+  .lostPassword {
+    font-size: 13px;
+    line-height: 20px;
+    a {
+      color: black;
+    }
+  }
+
+  .loginBtn {
+    background-color: #000;
+    color: #fff;
+    width: 100%;
+    height: 40px;
+    position: absolute;
+    bottom: 0;
+    font-size: 15px;
+    border: none;
+  }
+
+}
